+++ online-interpreter/app/view.php
@@ -11,15 +11,15 @@
         color: #eee;
     }
     .code-font{
-        
+        font-family: "Consolas", "Monaco", "Lucida Console", "Liberation Mono", "DejaVu Sans Mono", "Bitstream Vera Sans Mono", "Courier New";
     }
     </style>
 </head>
 <body>
     <div class="container">
-        <h1>Pashmak programming language online interpreter</h1>
+        <h1>Pashmak programming Language Online Interpreter</h1>
         <form method="POST" class="form-group">
-            <textarea style="background-color: #555; color: #eee; min-height: 300px;" class="form-control" placeholder="write your code..." name="code"><?= htmlspecialchars($code) ?></textarea>
+            <textarea style="background-color: #555; color: #eee; min-height: 300px;" class="form-control code-font" placeholder="write your code..." name="code"><?= htmlspecialchars($code) ?></textarea>
             <textarea style="background-color: #555; color: #eee; min-height: 100px;" class="form-control" placeholder="program stdin inputs" name="stdin"><?= htmlspecialchars($stdin) ?></textarea>
             <input class="btn btn-dark" style="width: 100%; margin-top: 10px;" type="submit" value="run" />
         </form>
@@ -32,7 +32,7 @@
         <a href="https://github.com/parsampsh/pashmak/blob/master/README.md#documentation" target="_blank">Pashmak Documentation</a>
         <br />
         <span style="color: gray;">sorry for website bullshit design, i'm not designer</span>
-        <br />
     </div>
+    <br /><br /><br />
 </body>
 </html>