+++ tests/items/test_set.py
@@ -68,3 +68,15 @@ class test_set(TestCore):
             $myvar = 'num is ' + str(2*5);
             out $myvar;
         '''), 'num is 10')
+
+        self.assert_output(self.run_without_error('''
+            mem 'hello world';
+            $myvar = ^;
+            out $myvar;
+        '''), 'hello world')
+
+        self.assert_output(self.run_without_error('''
+            mem 'hello world';
+            $myvar =    ^ ;
+            out $myvar;
+        '''), 'hello world')